# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample gin server.",
        "title": "Swagger Example API",
        "contact": {
            "name": "khr",
            "email": "hengruikou@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:9999",
    "basePath": "/api",
    "paths": {
        "/auth/info": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "详情接口",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/comDto.ResponseData"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "登录接口",
                "parameters": [
                    {
                        "description": "参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/reqDto.AdminLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/comDto.ResponseData"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "登出接口",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/captcha": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "summary": "获取验证码",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/comDto.ResponseData"
                        }
                    }
                }
            }
        },
        "/download": {
            "get": {
                "produces": [
                    "multipart/form-data"
                ],
                "tags": [
                    "public"
                ],
                "summary": "文件流下载",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/comDto.ResponseData"
                        }
                    }
                }
            }
        },
        "/upload/file": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "multipart/form-data",
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "summary": "上传单个文件",
                "parameters": [
                    {
                        "type": "file",
                        "description": "File to be uploaded",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/comDto.ResponseData"
                        }
                    }
                }
            }
        },
        "/upload/files": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "multipart/form-data",
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "summary": "上传视频",
                "parameters": [
                    {
                        "type": "file",
                        "description": "File to be uploaded",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/comDto.ResponseData"
                        }
                    }
                }
            }
        },
        "/uploads/files": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "multipart/form-data",
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "summary": "上传多个文件",
                "parameters": [
                    {
                        "type": "file",
                        "description": "File to be uploaded",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/comDto.ResponseData"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "comDto.ResponseData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "reqDto.AdminLogin": {
            "type": "object",
            "required": [
                "code",
                "method",
                "password",
                "uuid"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "method": {
                    "type": "string",
                    "enum": [
                        "name",
                        "phone"
                    ],
                    "example": "name"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "description": "binding:\"len=12\"",
                    "type": "string"
                },
                "revoke": {
                    "description": "binding:\"required\"",
                    "type": "boolean",
                    "example": false
                },
                "user_name": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        }
    },
    "tags": [
        {
            "description": "This is a test tag",
            "name": "TestTag1",
            "externalDocs": {
                "description": "This is my blog site"
            }
        }
    ],
    "x-example-key": {
        "key": "value"
    }
}